Authorization, Policies, Gates

+++ app/Http/Controllers/PostsController.php
@@ -2,18 +2,28 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\BlogPost;
+use App\BlogPost;
 use App\Http\Requests\StorePost;
 use Illuminate\Http\Request;
 
 // use Illuminate\Support\Facades\DB;
 
-class PostsController extends Controller
+// [
+//     'show' => 'view',
+//     'create' => 'create',
+//     'store' => 'create',
+//     'edit' => 'update',
+//     'update' => 'update',
+//     'destroy' => 'delete',
+// ]
+class PostController extends Controller
 {
-    public function __construct(){
+    public function __construct()
+    {
         $this->middleware('auth')
-            ->only(['create', 'store', 'update', 'edit', 'destroy']);
+            ->only(['create', 'store', 'edit', 'update', 'destroy']);
     }
+
     /**
      * Display a listing of the resource.
      *
@@ -51,12 +61,13 @@ class PostsController extends Controller
     public function show($id)
     {
         return view('posts.show', [
-            'post' => BlogPost::with('comments')->findOrFail($id)
+            'post' => BlogPost::with('comments')->findOrFail($id),
         ]);
     }
 
     public function create()
     {
+        // $this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -72,12 +83,20 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+        $this->authorize($post);
+
         return view('posts.edit', ['post' => $post]);
     }
 
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post!");
+        // }
+        $this->authorize($post);
+
         $validatedData = $request->validated();
 
         $post->fill($validatedData);
@@ -90,6 +109,12 @@ class PostsController extends Controller
     public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post!");
+        // }
+        $this->authorize($post);
+
         $post->delete();
 
         // BlogPost::destroy($id);
